requirements.txt
# Core Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-WTF==1.2.1

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Password Hashing
Werkzeug==3.0.1

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23

# Document Conversion (for development - Windows/Mac)
docx2pdf==0.1.8

# File Handling
python-magic-bin==0.4.14

# Forms & Validation
email-validator==2.1.0.post1

# Environment Variables
python-dotenv==1.0.0

# Security
itsdangerous==2.1.2

# Development
pytest==7.4.3
pytest-flask==1.3.0
